Начат рефакторинг модулей в MVP
